+++ src/DTO/BaseDTO.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Lloricode\LaravelPandagoSdk\DTO;
+
+use Spatie\DataTransferObject\DataTransferObject;
+
+abstract class BaseDTO extends DataTransferObject
+{
+    protected bool $ignoreMissing = true;
+}