+++ src/DTO/Order/DriverDTO.php
@@ -2,9 +2,9 @@
 
 namespace Lloricode\LaravelPandagoSdk\DTO\Order;
 
-use Spatie\DataTransferObject\DataTransferObject;
+use Lloricode\LaravelPandagoSdk\DTO\BaseDTO;
 
-class DriverDTO extends DataTransferObject
+class DriverDTO extends BaseDTO
 {
     public ?string $id;
     public ?string $name;