+++ src/DTO/Order/LocationDTO.php
@@ -2,9 +2,9 @@
 
 namespace Lloricode\LaravelPandagoSdk\DTO\Order;
 
-use Spatie\DataTransferObject\DataTransferObject;
+use Lloricode\LaravelPandagoSdk\DTO\BaseDTO;
 
-class LocationDTO extends DataTransferObject
+class LocationDTO extends BaseDTO
 {
     public string $address;
     public float $latitude;