+++ src/DTO/Order/OrderDTO.php
@@ -2,10 +2,10 @@
 
 namespace Lloricode\LaravelPandagoSdk\DTO\Order;
 
+use Lloricode\LaravelPandagoSdk\DTO\BaseDTO;
 use Lloricode\LaravelPandagoSdk\Enum\PaymentMethod;
-use Spatie\DataTransferObject\DataTransferObject;
 
-class OrderDTO extends DataTransferObject
+class OrderDTO extends BaseDTO
 {
     public SenderDTO $sender;
     public RecipientDTO $recipient;