+++ src/DTO/Order/RecipientDTO.php
@@ -2,9 +2,9 @@
 
 namespace Lloricode\LaravelPandagoSdk\DTO\Order;
 
-use Spatie\DataTransferObject\DataTransferObject;
+use Lloricode\LaravelPandagoSdk\DTO\BaseDTO;
 
-class RecipientDTO extends DataTransferObject
+class RecipientDTO extends BaseDTO
 {
     public string $name;
     public string $phone_number;